+++ __mocks__/fusion:properties.js
@@ -2,5 +2,5 @@ export default jest.fn().mockReturnValue({
   resizerURL: 'hi',
   feedDomainURL: 'http://demo-prod.origin.arcpublishing.com',
   feedTitle: 'google news',
-  feedLanguage: 'en'
+  feedLanguage: 'en',
 })