+++ blocks/sitemap-video-feature-block/searchHelper.js
@@ -1,0 +1,10 @@
+function formatSearchObject(searchObject) {
+  const keys = Object.keys(searchObject)
+  const jsmSearcher = keys.reduce((acc, key) => {
+    return [...acc, `${key} == \`${searchObject[key]}\``]
+  }, [])
+
+  return jsmSearcher
+}
+
+export default formatSearchObject