+++ blocks/feeds-source-content-api-block/CHANGELOG.md
@@ -1,8 +1,23 @@
 # @wpmedia/feeds-source-content-api-block
 
-## 0.1.0
+## 0.2.1
+
+### Patch Changes
+
+- [`779b521`](https://github.com/WPMedia/feed-components/commit/779b5214fc473832cd972f1cdff6d54c9500a90b) [#37](https://github.com/WPMedia/feed-components/pull/37) Thanks [@camdub](https://github.com/camdub)! - add content base property to ensure valid uri
+
+## 0.2.0
+
 ### Minor Changes
 
+- [`bd1fd38`](https://github.com/WPMedia/feed-components/commit/bd1fd38d1799173fa829e57782ddc9675b2d6f78) [#30](https://github.com/WPMedia/feed-components/pull/30) Thanks [@ibeyene](https://github.com/ibeyene)! - Updated default query parameters
 
+### Patch Changes
+
+- [`bd1fd38`](https://github.com/WPMedia/feed-components/commit/bd1fd38d1799173fa829e57782ddc9675b2d6f78) [#30](https://github.com/WPMedia/feed-components/pull/30) Thanks [@ibeyene](https://github.com/ibeyene)! - Add a google news sitemap to a package and cleanup work for other packages.
+
+## 0.1.0
+
+### Minor Changes
 
 - [`3f48092`](https://github.com/WPMedia/feed-components/commit/3f480923378341a9d5dd56c905c91398587c1135) Thanks [@camdub](https://github.com/camdub)! - Update all packages to autopublish