+++ blocks/sitemaps-feature-block/CHANGELOG.md
@@ -1,8 +1,13 @@
 # @wpmedia/sitemaps-feature-block
 
-## 0.1.0
+## 0.2.0
+
 ### Minor Changes
 
+- [`bd1fd38`](https://github.com/WPMedia/feed-components/commit/bd1fd38d1799173fa829e57782ddc9675b2d6f78) [#30](https://github.com/WPMedia/feed-components/pull/30) Thanks [@ibeyene](https://github.com/ibeyene)! - Fixed changefreq
 
+## 0.1.0
+
+### Minor Changes
 
 - [`3f48092`](https://github.com/WPMedia/feed-components/commit/3f480923378341a9d5dd56c905c91398587c1135) Thanks [@camdub](https://github.com/camdub)! - Update all packages to autopublish